Calculate challenges.rel.12.25

+++ lib/controllers/routine_user_controller.dart
@@ -53,7 +53,7 @@ class RoutineUserController extends ChangeNotifier {
     }
   }
 
-  int weight() {
+  double weight() {
     return _amplifyRoutineUserRepository.user?.weight ?? 0;
   }
